added unittests for some plugins

+++ plugin_examples/audacity_keys.py
@@ -1,15 +1,11 @@
 """HotKeys plugin for Audacity
 """
 import os.path
-## import sys
 import shutil
 import xml.etree.ElementTree as ET
-import collections
-## import functools
-## import string
 from ..gui import show_cancel_message, get_file_to_open, get_file_to_save
 
-instructions = """\
+INSTRUCTIONS = """\
 Instructions for rebuilding the key binding definitions
 
 
@@ -40,27 +36,29 @@ def build_data(page, showinfo=True):
     """lees de keyboard definities uit het/de settings file(s) van het tool zelf
     en geef ze terug voor schrijven naar het keydef bestand
     """
-    shortcuts = collections.OrderedDict()
     otherstuff = {}
-
-    try:
-        initial = page.settings['AC_KEYS']
-    except KeyError:
-        initial = ''
+    initial = page.settings.get('AC_KEYS', '')
 
     kbfile = ''
     if showinfo:
-        ok = show_cancel_message(page.gui, text=instructions)
+        ok = show_cancel_message(page.gui, text=INSTRUCTIONS)
         if ok:
             kbfile = get_file_to_open(page.gui, extension='XML files (*.xml)', start=initial)
+            # if kbfile:
+            #     page.settings['AC_KEYS'] = kbfile
     else:
         kbfile = initial
     if not kbfile:
-        return [], []
+        return {}, {}
 
     tree = ET.parse(kbfile)
-    root = tree.getroot()
-    ## data = []
+    shortcuts, otherstuff['commands'] = build_commandlist(tree.getroot())
+    return shortcuts, otherstuff
+
+
+def build_commandlist(root):
+    "build the list of commands"
+    shortcuts = {}
     key = 0
     commandlist = {}
     for item in root.findall('command'):
@@ -81,8 +79,7 @@ def build_data(page, showinfo=True):
             key += 1
             shortcuts[key] = (_translate_keyname(keyname), keymods, cmd_name, cmd_label)
         commandlist[cmd_name] = cmd_label
-    otherstuff['commands'] = commandlist
-    return shortcuts, otherstuff
+    return shortcuts, commandlist
 
 
 how_to_save = """\
@@ -107,11 +104,11 @@ def savekeys(parent):
     if not ok:
         return
 
-    try:
-        kbfile = parent.settings['AC_KEYS']
-    except KeyError:
+    kbfile = parent.settings.get('AC_KEYS', '')
+    if not kbfile:
         initial = os.path.expanduser('~/.config/Audacity-keys.xml')
         kbfile = get_file_to_save(parent, extension='XML files (*.xml)', start=initial)
+        parent.settings['AC_KEYS'] = kbfile
 
     root = ET.Element('audacitykeyboard')
     root.set('audacityversion', "2.0.5")
@@ -127,7 +124,8 @@ def savekeys(parent):
             key = 'Ctrl+' + key
         new.set('key', key)
 
-    shutil.copyfile(kbfile, kbfile + '.bak')
+    if os.path.exists(kbfile):
+        shutil.copyfile(kbfile, kbfile + '.bak')
     ET.ElementTree(root).write(kbfile, encoding="UTF-8")
